+++ CHANGELOG.md
@@ -7,5 +7,9 @@ This release focuses on a simplification of the configuration; a seemingly every
 #### New
 
 - Added the Particle layer type to the Aura Watcher Designer
+- Added a button to toggle a new appearance frame to the Interface Designer
+- Added a button to toggle the configuration interface from within the Interface Designer
+- Added a button to access the Aura Designer from within the Interface Designer
+- Cleaned up the Cast Bar designer so its easier to understand
 
 #### Other